Only allow pushing a single object at a time with a bias towards the closer object.
I've noticed that it was possible to push more than one rock at a time which can lead to accidentally pushing something you didn't mean to push. Two places I've seen this happen now is the forest cave with the heart piece and skulls, and the cave to enter Tal Tal Mountain Range where you have to move a few rocks to get to the stairway. I even accidentally pushed it over the stairs the one time, which is another bug for another time....