Another "hookshot is stuck and wont' fire" fix, except this time it's fool proof.
A little while ago I found the hookshot could get stuck if interrupting it with another weapon. This was really easy to do, so I implemented a quick fix for it: https://github.com/BigheadSMZ/Zelda-LA-DX-HD-Updated/commit/a95a92c9109d393c5e9b66abffb28abce4f4efa1

Looks like there are other ways to get stuck, like pulling in a piece of power or acorn at close range. The hookshot is instantly despawned and is never "returned". So to fix this once and for all, a countdown is used to check if the hookshot has returned yet. If it hasn't, reset the firing variable. Easy.