Fix Slime Eel death sound effects and music cues & extend AI damage state to skip playing "built-in" death sound.
The systems in place tried to automate some boss deaths, very few use it and I'm not sure why. But these bosses are poorly done due to this system. Missing "hum" after death, bad timing on death sound, missing sound effects, etc. Instead of relying on it for non-existent sounds and cues, we make our own. The AI damage state for boss is still useful, just not for playing things when we need them.