Maye fix cave on path to level 7 that is open when it should need to be bombed?
I don't know if it's fixed, or if it was even broken, but every time I come through here, the wall was already open. I checked the walls and the tiles had the wrong rotation so I fixed it. I came here with a new save by warping, and it appears to be fine now. So, maybe its fixed?