Reduce final boss "Giant Zol" chance to despawn to 33% down from 50%.
This was rather irritating. I may be wrong, but I don't remember it despawning quite so often on the original games. Even at 33% It's still a 1 in 3 chance to despawn so it can still happen, but at least it's much less likely it will happen like 3-5 times in a row like it could.