Set Kanalet Castle upstairs button to "pressed" state when finishing level 2.
I forgot about this when fixing these buttons. At some point in the game before collecting the leaves, the upstairs button should be set to a "pressed" state when first encountered if the button downstairs wasn't pressed. Finishing level 2 is a good place to put this since it's right before the golden leaves quest. Currently pressing it will do nothing if not pressing the downstairs button first since the gate is already closed.