Fix 2x Like Like bugs: steal shield from L/R & zero velocity when boots running.
Like Like wouldn't steal a shield from the player if the shield was on the top buttons as opposed to the face buttons. This is my bad, when adding two more inventory slots I overlooked that these slots were checked in even more places. Another bug was when boots running through a Like Like without hitting it with sword or shield, the player sprite would disappear but zip past the Like Like, which lead to an invisible Link leaving a dust trail but still technically "trapped".