Add sound effects when opening/navigating the island map.
Simple change. The only tricky part was the sound effect when opening the map since it doesn't actually track when the map is actually visible over the inventory. Or it does and I just don't get it. Either way, I added my own bool to know when the map is opened or closed since it's more straightforward to read.