Create more robust rotation system when under influence of vacuums (levels 2 & 8).
There was plenty of bugs with link rotating when the vacuum was pulling him in, and it was quite difficult to get right. Basically there was just a lot more tracking put into place and an alternate "direction" specifically used when rotating. This way the standard direction can still influence the logic while the rotational direction affects Link's current facing.