Move logic for replacing placeholder tags with glyphs to Language Manager.
This moves the logic for replacing placeholder tags like "[RIGHT]" with the glyph that translates into the font character from the font sheet (in this case an arrow pointing right using the glyph "±"). Before placeholder tags could only be used in "dialog" language strings (strings in "dialog_*.lng"), now it also applies to strings used for main menu elements (*.lng).