Entering Level 8 backdoors trigger save position to main entrance.
The two back exits in Level 8 are a weird one. When leaving to the overworld, the game is supposed to act as if the player never left the dungeon. Along the outside path is a world teleporter. Teleporting to this location saves your position. When entering one of the two Level 8 doors, the game does not autosave, but it should know you are now in Level 8. Saving, quitting, then loading the game used to put you back at the teleporter, but this hack makes it so you are properly moved to the Dungeon 8 entrance. This is also how it worked in the original games.