Add Totaka's Song Remix easter egg at file selection screen.
This was only present in the Japan and German versions, but since the song is in the "awakening.gbs" file, may as well make a way to access it. Plus its just fun to have these things. In the Japan version, it was triggered by entering "Totakeke" and in the German version, it was triggered by entering "MOYSE". I have also added an English equivalent by just entering "Totaka". The name is not case sensitive, so capitalization does not matter.