Do not load a save file with the "freezeGame" state set to "1".
Several situations in the game set the variable "freezeGame" to "1". This is used during various sequences to lock all objects in place during some scenes and actions. If the game was saved while this value was set to "1", then the save file would forever be tainted as nothing would ever be able to move again. So to prevent this, force a value of "0" as soon as a save file is loaded.