Update "photos.png" for all languages.
This fixes some of the icons, separates the button text from the buttons, and includes all of the missing text. None of this stuff is currently implemented. I plan to do something with it eventually, so at the very least I wanted everything to be available for when that time comes.