Add workaround for spawning inside doors when falling down holes.
This adds a workaround for when entering a room, the door closes, a hole is present, and there isn't already an "alternate hole respawn point" present. The original developers in many cases already had a fix for doors covering the respawn point, but it did not cover all cases. This workaround can serve to pick up the slack for those cases.

This change also makes it so Link does not take damage while falling down a hole. Collision will still happen, but no damage or knockback will come from it.