NPC sword collision also checks for "IsActive" for unspawned NPCs
Another fun one. Apparently even though an NPC isn't "spawned" into the world, it still has collision! Even though the NPC is not there, it's there. So in my check I have to also check if the body "IsActive == false" or it will detect the body's collision types. This isn't automatically handled or something. Because why would it be, that would be far too convenient. No wonder there is so many "ands" and "ors" scattered everywhere and "ifs" and "elses"....