Create "ModFile.cs" with "ModFile" class to reduce code bloat.
Any object that can parse the standard "lahdmod" file has this method in it, so instead of copy and pasting it over and over may as well have a centralized location for it. This "Parse" method should work for the majority of current lahdmod files and a few that haven't been released yet.