Implement preliminary "Classic Camera" and move camera options to dedicated menu.
This is the beginning of a "Classic Camera" option that mimics the original "scrolling screen" camera system of the original games. This change also implements a camera settings menu and moves the other camera options here as well as implementing a few new ones pertaining to the classic camera. There is still a lot of fine tuning needed, so I figured I would implement this part now in a commit and then fix the rest of it in smaller commits. This commit already contains a bunch of stuff to update the menus, so I don't want to eventually have one giant hard to read commit.

Here is a summary of the changes in this commit:
- Create a "Camera Settings" menu.
- Move "Lock Camera" and "Smooth Camera" here.
- Implement the preliminary "Classic Camera".
- Add supporting options: Classic Border, Border Fill, and Fill Alpha.
- Organize code for the game settings variables.
- Update language files with new options.