Fix dungeon Horse Heads so they can't be thrown through the wall.
Because colliders have 16 units of depth, and Link holds the horse head above 16 units, it was possible to throw the horse head over the wall. Once collision has been breached, nothing will stop an object from continuing to move through the collision so it just kept going even when descending below 16 units. This commit just adds a second layer of "collision" by adding two additional collision boxes to the horse head that are checked every frame. When the two boxes meet the criteria for collision, the horse head will bounce back into the room by reversing it's velocity on the axis it was thrown. This simulates both "collision" and a "bouncing" effect.