Add "charge jumping" animations and play them when charge jumping.
This creates 12 new animations when jumping while charging. They are basic 1 frame animations that match the pose Link performs in the original games while jumping. While there was only 4 animations in the original, this adds 8 more due to being able to hold the shield while charge jumping which was not possible in the original game. There is 4 for the original shield and 4 for the mirror shield, though they are nearly identical.