Squeeze through Dungeon 1 doorway if pushing block towards door.
This removes 2 pixels worth of collision from the passageway between the room with the stairway to the feather and the room where a block is pushed to open it up. In the original game, it's possible to break collision just enough to squeeze through tight openings like this which is not possible with the type of collision this port uses. I don't think this fix should cause any problems.