Rework game scale to be a multiplier of original game scale and increase max to 20x.
This reworks the scale slider to go up to 20x and no longer scales by a resolution that doesn't make sense. The resolution of the GameBoy was 160x144, with the game field being 160x128. Scaling is now calculated from this base resolution which unlocks a lot more "in-between" scaling opportunities. Before the scale was calculated from a base resolution of 380x256, so the term "scale" didn't really mean anything in regards to the original game resolution.

 This also changes autoscaling behavior when adjusting the scale with the controller. When setting autoscaling with LT+RT, then adjusting the scale up or down, it will act as if you are adjusting from the equivalent scaling value instead of the "last used" scaling value before autoscale was set. This works even if the window size changes, which automatically adjusts the autoscale set. Overall, it makes scaling with the controller feel a lot "smoother".