Classic Camera: Gray out options that work with or without it.
The menu systems never had a way to show if some options affect others. For example in a Windows form where some options enable or disable others, the disabled options appear "grayed out". This is kind of the effect I was going for here. It doesn't prevent changing grayed out options, it's more of a way to show that the option currently has no effect.