Classic Camera: Implement "always animate" object group & add Vire.
This creates an object group to add game objects to that are always animated when Classic Camera is active and they are outside of the current field. The first objects to be added to this group are Vire, the bats that Vire spawns into, and Evil Eagle which previous had a dedicated value to always animate.