Set to v1.4.6 + rework patcher to store/extract zip file rather than patch files.
I can't tell you how many times Visual Studio has bugged out when adding patches and created a second resources file named "Resources1.resx" instead of using the existing "Resources.resx" file when adding the patch files. I have lost hours trying to fix it when it happens and have tried everything to prevent it from happening. I think it has something to do with deleting old patches and them not being removed from the "Resources.Designer.cs" file. When it fails to do this I have to manually go through and delete them from the list. I have followed all the tips online I could find, I've fixed it numerous times, and it keeps happening so I'm just done with it. It was a bad system to begin with.

The patcher now stores a single resource: a ZIP file that contains all the patches. It then writes it to disk and extracts them into the patches folder. This makes it much easier on me. If it bugs out when adding it, it's just one file. Adding it is fast, a drag and drop and it's done. This beats adding over 300 files one by one because Visual Studio is ass when it comes to importing multiple resources. Can't mass edit the "Neutral Comment", can't mass import all files at once, drag and drop and hit enter 300+ times. Well no more. Using a ZIP file makes everything way easier and makes it far less daunting to create a new patch. Not sure why I didn't do it this way to begin with since I've done it this way with my projects in the past like Custom Texture Tool PS.