Remove the "FreezeWorldAroundPlayer" value from GameManager.
I never really knew what it did or what it was supposed to do. But at this point, it really does nothing. The one thing it was there for was when picking something up. Maybe it was supposed to freeze enemies during a pickup? This would make sense for Spiny Beetles maybe, which in theory would freeze them while lifting the object off their back. But it didn't do that, it didn't seem to do anything. Removing this doesn't seem to have any ill effects on picking up objects, plus the new pickup code makes picking up objects uninterruptible so there really is no reason to keep it around and adding more overhead to the update loop.