Classic Camera: Respawn bushes/grass on a field change.
The original game respawned shrubbery each time the screen scrolled, so let's do that here too. This uses the "FieldChange" variable on "ObjLink" to know when to respawn the grass/bush objects. There is a slight delay so the respawn isn't visible until after the camera pans, but it's probably still possible to view it if the player is quick enough.