Recreate Dungeon 3 as a single map + update LADXHD tools.
Dungeon 3 was split up into 4 maps for no good reason. This recreates the entire dungeon as a single map like the original game. When it was split up into 4 separate maps, the minimap was completely wrong. It's supposed to look like two keys side by side: a large and a small one. Because of the name change to the map, the migration tool, patcher, and patch generation scripts all needed to be updated to handle the new file. The overworld also needed an update to reference the new map name.