Fix weird bug with Turtle Rock miniboss by resetting while it attacks.
I missed a single crucial value: "_body.VelocityTarget". When this is set, the "body" component will try to meet that value. Since it was no longer being updated afterwards, whatever the velocity was when resetting, it was stuck at that velocity, so the target was never met. Well whatever, also reset Velocity just to be extra safe, and it turns out I missed the "_attackable" variable as well but I don't think it was nearly as important as the others.