Fix potential softlock in cave near Animal Village.
If a bomb arrow was used to get the heart to open the hookshot rock, then the player jumped down and had no bombs left, there would be no way to leave the cave. So I added a skull with a bomb in it so the player can at least bomb their way out.