Classic Camera: Reset hole map after field transition.
I was thinking of the best way to go about this, but eventually settled on resetting the entire hole map at once. In this port, digging holes is stored in a 2D array that is the size of the tilemap. There is a total of 160x128 tiles on the overworld. Referencing a tile means referencing the X and Y value. For example, "HoleMap.ArrayTileMap[80,112,0]" refers to the 80th tile from the left, and the 112th tile from the top. The hole map stores a "0" for the current tile reference when a hole is not dug, and stores a "-1" when a hole is dug. The holemap array is actually 3-dimensional, but the 3rd value is always zero (at least for holes).

This commit resets the entire tilemap when the field transitions. The reason is that it's just simpler than the alternative which is to reset the current field tiles. It would probably be more optimal to only reset the tiles within the current field, but the entire tilemap reset is both fast and asynchronously done, so it does not cause much if any performance penalty.