Fix egg follower turnaround when carrying rooster.
This used to be handled in script.zScript, but there isn't a way to throw an object using it to my knowledge. While I probably could have (and should have) implemented this action, I just decided to use Link's key change listener to perform the same actions. It took quite a fair bit more code to pull off, but the result is nearly identical with the added bonus of dropping the chicken.