Reduce Ghini fairy spawn rate to 10% and can drop other items.
The way Ghini drops worked was kind of weird. When killed it had a 75% chance to drop a fairy but no other items. This means there was always a 25% chance to drop rupees, hearts, and powerups, but it was hardly noticeable that other items did drop. So this reduces the fairy spawn rate to 10% and also doesn't prevent other items from dropping simultaneously, leading to increased loot from them.