Classic Camera: Fix final two egg paths that take you right before up.
This problem is difficult to explain but I'll try. There is one inherent difference between normal and classic cameras when running around the egg, and that is which rooms serves as the main rooms. There is 4 identical rooms that are cycled between when running around the egg to give the illusion of an endless maze. When teleported around, 2/4 of the rooms serve as the "main" rooms. The normal camera uses the bottom rooms, while the classic camera uses the top rooms. Due to how complex the maze logic is, this is just how it ended up when adding classic camera as it needed to use the top rooms for the scrolling illusion. Anyway, there was logic to complete the maze for normal camera but not classic when in the "second" set of rooms. Classic camera worked when in the "left-top" room because right above it is the exit. But it failed when in the "right-top room" because there is nothing above it.