Classic Camera: Lock onto camera field objects when exiting maps with them.
When camera field objects were introduced, it was made so that when entering a map with them, the camera will temporarily lock onto the closest one. After the transition is completed, the camera will then lock onto Link as normal to determine the current field. This makes the transition in always show the proper field, but it did not work for going in the opposite direction.

This commit makes it so when exiting an area with camera field objects, the camera will temporarily lock onto them until the next part of the transition which will either lock onto a new camera field object or will be a return to the overworld. This prevents the camera from ever moving into an "invalid field" or rather a field that is not part of the actual map.