Rework the boomerang so it can collect multiple items on one throw.
This commit creates and adds items to an item list when the boomerang comes in contact with them. The list is iterated over when updating item position and when the boomerang returns and collection starts. This does not allow collecting multiple fairies since they are technically not an "item" and require a separate loop. I did not want to take a chance with having far too many things happening at once though I doubt it would have been a problem.