Fix pushable stone quirks: push over door + breaking pushing.
This commit fixes two quirks with "ObjMoveStone". The first is that it was possible to push a stone over top of a passageway. To prevent this, I gave "ObjDoor" a new collision type and made it so pushable stones can't be pushed into it. Nothing else collides with this collision type. The second issue was a lot more serious where pushing could be broken completely. Sometimes when pushing a stone down a hole, then resetting stone positions with a "leave button", the stone would forever be stuck in a "moving" state which prevented other stones from being pushable. The new code does not rely on this behavior anymore so it has been removed.