Classic Camera: Always animate Marin and the Great Fairy.
The music changes when the fairy spawns or when Marin sings. Due to the way classic camera and these song changes work, switching fields does not change the song back to normal. Before the field was shrunk by 1 pixel on each edge of the screen, but it's possible to bypass this 1 pixel border by dashing with the Pegaus Boots. Link can blast past it before the next update loop hits.