Add a new "ObjPushButton" object that pushes Link forward.
This object was created to force Link to walk through a doorway. It will be implemented for boss fights in dungeons in future commits. Basically, Link steps on the tile, and it forces him to walk forward until exiting the allotted space. The button can be destroyed by setting the same key that a boss uses to store it's defeated state.