Another ObjButtonPush rework + add more of them.
This should hopefully be the final form of this thing. Instead of a "destroy" key, it now has an "active" key and the object is never actually destroyed. It uses the active key to know when it should push the player, so the required value is an inversion of what it was. So if a key was named "d1_et_1" the destroy key would be "1" to destroy it, now it needs "0" to remain active until the key is set which deactivates it.

This commit also adds many more "push buttons" to the various dungeons wherever a door closes behind the player. This is as always a way to try to mimic how the original game operates where doors close behind the player after a screen transition. Obviously this is slightly different with the modern camera, but it still pushes when its set as the camera.