Better handle ocarina teleport situations and hacks.
This simplifies the hack when teleporting to Manbo's pond, and it fixes using the ocarina in two specific dungeon situations to face Link in the correct direction. If playing Manbo's song on floors 2+ in Level 7 dungeon, the player would face south instead of north after the teleport. The same was true if entering Level 8 through a backdoor. This commit fixes both of these situations.