Add music trigger around Level 7 Hinox field rect.
This is the same method used to reset the music back to dungeon music when fighting Dodongo Snakes. This is not usually necessary, it was added as a fallback if a player manages to leave the field while running which has the possibility of skipping the trigger to restart the music from the fight.