Major "ObjLink" cleanups/refactoring.
This was a long time coming, something I've wanted to do since the beginning. This refactors and categorizes all the code on "ObjLink" so it's easier to navigate and understand. It divides the code into sections based on specific categories, which makes it easier to find all the code for that specific category.

This commit also adds a new variable to BodyComponent named "IgnoreCollsion" which can disable collsion for the body. This can be enabled by simply writing something like `_body.IgnoreCollision = true;` and the objects body will now be able to pass through all collision. This is really only useful for debugging purposes, at least I can't think of any other situations it could be useful at the moment.

One of the more notable changes is that a lot of miscellaneous code is moved out of "ObjLink" and into "ObjLinkHelper". The vast majority of the code moved is either "helper" functions or "one-offs" that either set or reference properties of "ObjLink" and have no other impact on overall functionality.