Classic Camera: Fix some things not respawning on field change.
This was a tricky one to figure out. For the longest time, there has been an issue where sometimes 1 or more objects would not respawn on a field change, usually bushes or grass. The issue was in "ObjectManager.cs" all along, setting "Link.FieldChange = false" prematurely would prevent some things to not respawn. The "FieldChange" did not need a manual reset to "false" since the next game loop would take care of that in ObjLink. So I think this ended up being some kind of race condition. FieldChange was going false before the game had a chance to respawn these objects that rely on it being "true". At least, that is my best guess.